fix creating and joining events
