+++ app/src/androidTest/java/space/collabify/android/unit/LoginScreenActivityTest.java
@@ -10,7 +10,7 @@ import android.widget.ImageButton;
 import space.collabify.android.R;
 import space.collabify.android.activities.LoginScreenActivity;
 import space.collabify.android.activities.ModeSelectActivity;
-import space.collabify.android.managers.AppManager2;
+import space.collabify.android.managers.AppManager;
 
 /**
  * This file was born on March 20, at 14:08
@@ -19,7 +19,7 @@ public class LoginScreenActivityTest extends ActivityInstrumentationTestCase2<Lo
 
     private ImageButton mLoginButton;
     private LoginScreenActivity mActivity;
-    private AppManager2 mAppManager;
+    private AppManager mAppManager;
 
     public LoginScreenActivityTest(){
         super(LoginScreenActivity.class);
@@ -33,7 +33,7 @@ public class LoginScreenActivityTest extends ActivityInstrumentationTestCase2<Lo
     protected void setUp() throws Exception {
         super.setUp();
         mActivity = getActivity();
-        mAppManager = AppManager2.getInstance();
+        mAppManager = AppManager.getInstance();
         mLoginButton = (ImageButton) mActivity.findViewById(R.id.loginButton);
     }
