+++ app/src/main/java/space/collabify/android/activities/JoinEventActivity.java
@@ -101,48 +101,52 @@ public class JoinEventActivity extends CollabifyActivity implements
 
 
     public void toCollabifier(Event event, String password){
-        //TODO may have to change how password is handled/displayed
-        if(event.getPassword().equalsIgnoreCase(password)){
-            mAppManager.getUser().setRole(Role.COLLABIFIER);
 
-            mAppManager.joinEvent(event.getId(), new CollabifyCallback<space.collabify.android.collabify.models.domain.User>() {
-                @Override
-                public void success(space.collabify.android.collabify.models.domain.User user, Response response) {
-                    User current = mAppManager.getUser();
-                    if (current.getId().equals(user.getUserId())) {
-                        Log.d(TAG, "Successfully joined");
-                        Intent intent = new Intent(JoinEventActivity.this, CollabifierActivity.class);
-                        startActivity(intent);
-                    }
-                }
-
-                @Override
-                public void failure(RetrofitError error) {
-                    Log.e(TAG, "Failed to join event:\n" + error.toString());
-
-                    runOnUiThread(new Runnable() {
-                        public void run() {
-                            Toast.makeText(JoinEventActivity.this, "Error joining Event. Please try again!", Toast.LENGTH_LONG).show();
-                        }
-                    });
-                }
-
-                @Override
-                public void exception(Exception e) {
-                    Log.e(TAG, "Failed to join event:\n" + e.getClass().getSimpleName());
-
-                    runOnUiThread(new Runnable() {
-                        public void run() {
-                            Toast.makeText(JoinEventActivity.this, "Error joining Event. Please try again!", Toast.LENGTH_LONG).show();
-                        }
-                    });
-                }
-            });
-
-        } else {
-            //bad password, don't do anything
-            Toast.makeText(JoinEventActivity.this, "Bad Password!", Toast.LENGTH_LONG).show();
+        if (event.isProtectedEvent()) {
+            if (event.getPassword() == null) {
+                Toast.makeText(JoinEventActivity.this, "There is an error with this event. Sorry :(", Toast.LENGTH_LONG).show();
+                return;
+            }
+            else if (!event.getPassword().equals(password)) {
+                Toast.makeText(JoinEventActivity.this, "Bad Password!", Toast.LENGTH_LONG).show();
+                return;
+            }
         }
+
+
+        mAppManager.joinEvent(event.getId(), new CollabifyCallback<space.collabify.android.collabify.models.domain.User>() {
+            @Override
+            public void success(space.collabify.android.collabify.models.domain.User user, Response response) {
+                User current = mAppManager.getUser();
+                if (current.getId().equals(user.getUserId())) {
+                    Log.d(TAG, "Successfully joined");
+                    Intent intent = new Intent(JoinEventActivity.this, CollabifierActivity.class);
+                    startActivity(intent);
+                }
+            }
+
+            @Override
+            public void failure(RetrofitError error) {
+                Log.e(TAG, "Failed to join event:\n" + error.toString());
+
+                runOnUiThread(new Runnable() {
+                    public void run() {
+                        Toast.makeText(JoinEventActivity.this, "Error joining Event. Please try again!", Toast.LENGTH_LONG).show();
+                    }
+                });
+            }
+
+            @Override
+            public void exception(Exception e) {
+                Log.e(TAG, "Failed to join event:\n" + e.getClass().getSimpleName());
+
+                runOnUiThread(new Runnable() {
+                    public void run() {
+                        Toast.makeText(JoinEventActivity.this, "Error joining Event. Please try again!", Toast.LENGTH_LONG).show();
+                    }
+                });
+            }
+        });
     }
 
     public void toCreateEvent(View v) {
